mock-tool/requirements.txt
flask
openai
python-dotenv
requests
httpx
